Omit max date limit on frequencies
By default, augur frequencies will use the most recent collection date
of the input data as its maximum date. This commit drops the explicit
max date argument used in the Nextstrain ncov workflow, so we get tip
frequencies that represent only the data in the tree without needing to
parameterize this max date elsewhere in the workflow.